Moved IV back to sets.py
